Completion of Save As, Save function and wiring of dialog box into parent form. Oops! - what about the error message for name?

+++ SQL/Environment Setup Scripts/Report Table Scripts.sql
@@ -18,7 +18,7 @@ CREATE TABLE `profitwisereport` (
     
     `CopyOfSystemReport` TINYINT NOT NULL,
     `CopyForEditing` TINYINT NOT NULL,
-    `OriginalReportId` TINYINT NULL,    
+    `OriginalReportId` BIGINT NULL,    
     
     `StartDate` DATE NOT NULL,
     `EndDate` DATE NOT NULL,
